api.js, challengeManager.js, formUtilities.js, gameManager.js created
